pyproject.toml
[project]
name = "duplocloud-docker-mcp"
version = "0.1.0"
description = "DuploCloud MCP server — exposes DuploCloud infrastructure management as tools consumable via Docker MCP Toolkit"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duplocloud-client>=0.4.0",
    "mcp>=1.26.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
